% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_dist_pop_data.R
\name{apply_growth_rate}
\alias{apply_growth_rate}
\title{Applies the growth rate to each population columns}
\usage{
apply_growth_rate(base_data, pop_column)
}
\arguments{
\item{base_data}{\code{tibble} Combined population data with the growth rate columns.}

\item{pop_column}{\code{str} Name of the population column to apply the growth rate to.}
}
\value{
\code{tibble} Population data with population filled based on growth rates
}
\description{
Applies the growth rate to each population columns
}
\keyword{internal}
