% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pop_data_quality.R
\name{compute_outliers_df}
\alias{compute_outliers_df}
\title{Checks whether outliers exist among age groups}
\usage{
compute_outliers_df(guid_age_data, guid_col)
}
\arguments{
\item{guid_age_data}{\code{tibble} Output of \link{filter_guid_age_data}.}
}
\value{
\code{tibble} Summary of outlier detection
}
\description{
Checks whether outliers exist among age groups
}
\keyword{internal}
