% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pop_data_quality.R
\name{create_agegroup_predicate}
\alias{create_agegroup_predicate}
\title{Create age group predicate per GUID for assertr validation}
\usage{
create_agegroup_predicate(guids_with_missing_agegroups)
}
\arguments{
\item{guids_with_missing_agegroups}{GUIDs with missing age groups.}
}
\value{
A predicate function for use with assertr.
}
\description{
Predicate that checks whether each GUID has required age group data.
}
\keyword{internal}
