% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dist_pop_map.R
\name{create_dist_pop_map}
\alias{create_dist_pop_map}
\title{Map district population}
\usage{
create_dist_pop_map(district_long, dist_pop, who_region, .year, country = NULL)
}
\arguments{
\item{district_long}{\code{sf} District shapefile in long format.}

\item{dist_pop}{\code{tibble} District population.}

\item{who_region}{\code{str} Region to map.}

\item{.year}{\code{int} Year to map.}

\item{country}{\code{str} Country to map (optional).}
}
\value{
\code{ggplot} A map with U15pop data.
}
\description{
The function acts as a diagnostic tool be quickly checking which districts are
missing population data.
}
