% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_prov_pop_map.R
\name{create_prov_pop_map}
\alias{create_prov_pop_map}
\title{Map province population}
\usage{
create_prov_pop_map(prov_long, prov_pop, who_region, .year, country = NULL)
}
\arguments{
\item{prov_long}{\code{sf} Province shapefile in long (year) format.}

\item{prov_pop}{\code{tibble} Cleaned province population.}

\item{who_region}{\code{str} Region to map.}

\item{.year}{\code{int} Year to map.}

\item{country}{\code{str} Country to map (optional).}
}
\value{
\code{ggplot} A map with U15pop data.
}
\description{
The function acts as a diagnostic tool be quickly checking which provinces are
missing population data.
}
