% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk_pop_cols.R
\name{crosswalk_pop_cols}
\alias{crosswalk_pop_cols}
\title{Crosswalks column names of the POLIS pop dataset}
\usage{
crosswalk_pop_cols(pop_data)
}
\arguments{
\item{pop_data}{\code{tibble} The population dataset from the POLIS API. It can be
the entire table or a subset of it.}
}
\value{
\code{tibble} A tibble with the renamed columns.
}
\description{
Crosswalks the original column names of the POLIS pop dataset so that it is
compatible with the names used by the data cleaning scripts developed within
CDC. The function also filters for the only columns used in the population
datasets and adds new columns, such as the adm{0-2}guids. In addition, the
tables get transformed by pivoting wider using the AgeGroupCode.
}
\examples{
\dontrun{
dist_pop <- load_polis_pop("dist")
dist_pop <- crosswalk_pop_cols(dist_pop)

}
}
