% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pop_data_quality.R
\name{detect_pop_outliers}
\alias{detect_pop_outliers}
\title{Detect population outliers}
\usage{
detect_pop_outliers(pop_data, guid_col, age_groups = c("0-5Y", "0-15Y", "ALL"))
}
\arguments{
\item{pop_data}{Tibble containing population data.}

\item{guid_col}{String name of the GUID column to validate.}

\item{age_groups}{\code{str} Age groups to analyze. By default, analyzes for the
under 5, under 15, and total population.}
}
\value{
\code{list} A list containing outliers per age group analyzed.
}
\description{
Detects the outliers in population counts for each age group and GUID.
}
\keyword{internal}
