% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_processed_pop_outliers.R
\name{detect_processed_pop_outliers}
\alias{detect_processed_pop_outliers}
\title{Detects whether there are issues with the population data}
\usage{
detect_processed_pop_outliers(cleaned_pop_data, group_col = "adm1guid")
}
\arguments{
\item{cleaned_pop_data}{\code{tibble} Cleaned population data. Either dist, prov, or country.}

\item{group_col}{\code{str} Name of the grouping variable. Recommend adm0guid, adm1guid, or adm2guid.}
}
\value{
\code{tibble} Pop data with category of outlier.
}
\description{
Detects whether there are issues with the population data
}
