% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_pop_data.R
\name{dist_pop_data}
\alias{dist_pop_data}
\title{Build district population (Admin2) in wide format}
\usage{
dist_pop_data(
  pop_data,
  dist_file_loc = "GID/PEB/SIR/Data/spatial/global.dist.rds",
  growth_rate_file_loc =
    "GID/PEB/SIR/Data/pop/pop raw/WPP2024_GEN_F01_DEMOGRAPHIC_INDICATORS_COMPACT.xlsx",
  output_file = getwd(),
  edav = TRUE
)
}
\arguments{
\item{pop_data}{Tibble POLIS datasource}

\item{output_file}{Optional .rds output path.}
}
\value{
Tibble with GUID + district-year rows and final output naming.
}
\description{
Combines POLIS + patches + Jamal; joins to district-year shapes; deduplicates;
fills datasource across gaps within GUID; fills missing values using growth rates.
}
\examples{
\dontrun{
dist_pop_data(pop_data, output_file = "Data/pop/dist_pop_admin2.rds")
}
}
