% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pop_data_quality.R
\name{filter_guid_age_data}
\alias{filter_guid_age_data}
\title{Filters data to a single (guid, age_group) slice and adds an index}
\usage{
filter_guid_age_data(pop_data, guid_col, guid, age_group)
}
\arguments{
\item{pop_data}{Tibble containing population data.}

\item{guid_col}{String name of the GUID column to validate.}

\item{guid}{\code{str} GUID to search for.}

\item{age_group}{\code{str} Age group to search for.}
}
\value{
\code{str} Pop data filtered to a particular GUID and age group
}
\description{
Filters data to a single (guid, age_group) slice and adds an index
}
\keyword{internal}
