% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pop_data_quality.R
\name{find_conflicting_admins}
\alias{find_conflicting_admins}
\title{Find conflicting admin mappings}
\usage{
find_conflicting_admins(
  pop_data,
  child_admin_col,
  parent_admin_col,
  return_details = FALSE
)
}
\arguments{
\item{pop_data}{\code{tibble} Population dataset.}

\item{child_admin_col}{\code{str} Column name of the next lower admin column. For example,
"Admin2GUID" if \code{pop_data} contains the province population dataset.}

\item{parent_admin_col}{\code{str} Parent admin column. For example, "Admin1GUID"
if \code{pop_data} contains the province population dataset.}

\item{return_details}{\code{logical}
If TRUE, returns a tibble with conflicting parent GUIDs per child.
If FALSE, returns the child GUIDs with multiple parents.}
}
\value{
\code{tibble} Summary tibble with conflict details or vector of conflicting GUIDs.
}
\description{
Identifies child admin units that map to multiple parent admin units.
}
\keyword{internals}
