% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pop_data_quality.R
\name{get_guid_missing_age_groups}
\alias{get_guid_missing_age_groups}
\title{Obtains GUID with missing age group}
\usage{
get_guid_missing_age_groups(pop_data, guid_col)
}
\arguments{
\item{pop_data}{Tibble containing population data.}

\item{guid_col}{String name of the GUID column to validate.}
}
\value{
\code{tibble} A summary table of GUIDs and their missing age groups.
}
\description{
Obtains the GUIDs missing age groups.
}
\keyword{internal}
