% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_pop_data.R
\name{load_all_patches}
\alias{load_all_patches}
\title{Combine all district patch sources}
\usage{
load_all_patches(
  pakistan_file_path =
    "GID/PEB/SIR/Data/pop/pop raw/csv files/2022_2023 Population Pakistan.csv",
  somalia_2022_file_path = "GID/PEB/SIR/Data/pop/pop raw/csv files/AFPPOP_22.csv",
  somalia_2023_file_path = "GID/PEB/SIR/Data/pop/pop raw/csv files/AFPPOP_23.csv",
  somalia_2024_file_path = "GID/PEB/SIR/Data/pop/pop raw/csv files/AFPPOP_24.csv",
  kenya_file_path = "GID/PEB/SIR/Data/pop/pop raw/csv files/Kenya_SubCounty_pop_2018.csv",
  jamal_pop_file_path = "GID/PEB/SIR/Data/pop/pop raw/csv files/POPU15.csv",
  edav = TRUE
)
}
\arguments{
\item{pakistan_file_path}{Path to Pakistan patch file.}

\item{somalia_2022_file_path}{Path to Somalia 2022 file.}

\item{somalia_2023_file_path}{Path to Somalia 2023 file.}

\item{somalia_2024_file_path}{Path to Somalia 2024 file.}

\item{kenya_file_path}{Path to Kenya 2018 file.}

\item{jamal_pop_file_path}{Path to Jamal file.}
}
\value{
\code{tibble} Dataset with all the patch files.
}
\description{
Combine all district patch sources
}
