% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_prov_pop_data.R
\name{load_dist_pop_rollup}
\alias{load_dist_pop_rollup}
\title{Calculate province level population using district pop rollups}
\usage{
load_dist_pop_rollup(
  dist_pop_file_path = file.path("GID/PEB/SIR/Data/pop", "processed_pop_file",
    "global.dist.parquet"),
  edav = TRUE
)
}
\arguments{
\item{dist_pop_file_path}{\code{str} Path to the district pop file.}

\item{edav}{\code{logical} If the file is in EDAV. Defaults to \code{TRUE}.}
}
\value{
\code{tibble} District population rollup.
}
\description{
Calculate province level population using district pop rollups
}
