% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_pop_data.R
\name{load_district_long}
\alias{load_district_long}
\title{Load district spatial table (district-year long)}
\usage{
load_district_long()
}
\value{
Tibble district long table.
}
\description{
Load district spatial table (district-year long)
}
