% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_pop_data.R
\name{load_growth_rates}
\alias{load_growth_rates}
\title{Load population growth rates}
\usage{
load_growth_rates(
  file_loc =
    "GID/PEB/SIR/Data/pop/pop raw/WPP2024_GEN_F01_DEMOGRAPHIC_INDICATORS_COMPACT.xlsx",
  edav = TRUE
)
}
\arguments{
\item{file_loc}{Path to WPP Excel file.}
}
\value{
Tibble with Admin0Name, year, growth_rate.
}
\description{
Load population growth rates
}
