% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_prov_pop_data.R
\name{load_indonesia_patch}
\alias{load_indonesia_patch}
\title{Indonesia province patch}
\usage{
load_indonesia_patch(
  indonesia_file_path =
    "GID/PEB/SIR/Data/pop/pop raw/csv files/Papua Population u15 2023.csv",
  edav = TRUE
)
}
\arguments{
\item{indonesia_file_path}{\code{str} File path to the Indonesia patch for 2023.}

\item{edav}{\code{logical} Whether the file is in EDAV.}
}
\value{
\code{tibble} Indonesia patch
}
\description{
Indonesia province patch
}
