% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_pop_data.R
\name{load_jamal_pop}
\alias{load_jamal_pop}
\title{Load Jamal district under-15 population}
\usage{
load_jamal_pop(
  file_loc = "GID/PEB/SIR/Data/pop/pop raw/csv files/POPU15.csv",
  edav = TRUE
)
}
\arguments{
\item{file_loc}{Path to Jamal file.}
}
\value{
Tibble with Admin names, year, age columns, datasource.
}
\description{
Load Jamal district under-15 population
}
