% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_dist_pop_data.R
\name{load_kenya_patch}
\alias{load_kenya_patch}
\title{Load Kenya district population patch}
\usage{
load_kenya_patch(
  kenya_file_path = "GID/PEB/SIR/Data/pop/pop raw/csv files/Kenya_SubCounty_pop_2018.csv",
  edav = TRUE
)
}
\arguments{
\item{kenya_file_path}{Path to Kenya 2018 file.}
}
\value{
Tibble with Admin names, year, age columns, datasource.
}
\description{
Load Kenya district population patch
}
