% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_pop_data.R
\name{load_pakistan_patch}
\alias{load_pakistan_patch}
\title{Load Pakistan district population patch}
\usage{
load_pakistan_patch(
  file_loc = "GID/PEB/SIR/Data/pop/pop raw/csv files/2022_2023 Population Pakistan.csv",
  edav = TRUE
)
}
\arguments{
\item{file_loc}{Path to patch file.}
}
\value{
Tibble with Admin names, year, age columns, datasource.
}
\description{
Load Pakistan district population patch
}
