% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_polis_pop.R
\name{load_polis_pop}
\alias{load_polis_pop}
\title{Load POLIS Population Data}
\usage{
load_polis_pop(
  spatial_scale,
  edav = TRUE,
  file_loc = "GID/PEB/SIR/POLIS/data/pop.rds",
  azcontainer = sirfunctions::get_azure_storage_connection()
)
}
\arguments{
\item{spatial_scale}{Geographic level: "ctry", "prov", or "dist"}

\item{edav}{Load data from EDAV}

\item{file_loc}{Load from local file (path)}

\item{azcontainer}{Azure storage connection}
}
\value{
Tibble filtered to spatial_scale with datasource column
}
\description{
Load POLIS Population Data
}
\examples{
\dontrun{
load_polis_pop(spatial_scale = "dist")
load_polis_pop(spatial_scale = "prov", file_loc = "path/to/file.rds")
}

}
