% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_pop_data.R
\name{load_somalia_patch}
\alias{load_somalia_patch}
\title{Load Somalia district population patch}
\usage{
load_somalia_patch(
  file_2022 = "GID/PEB/SIR/Data/pop/pop raw/csv files/AFPPOP_22.csv",
  file_2023 = "GID/PEB/SIR/Data/pop/pop raw/csv files/AFPPOP_23.csv",
  file_2024 = "GID/PEB/SIR/Data/pop/pop raw/csv files/AFPPOP_24.csv",
  edav = TRUE
)
}
\arguments{
\item{file_2022}{Path to 2022 file.}

\item{file_2023}{Path to 2023 file.}

\item{file_2024}{Path to 2024 file.}
}
\value{
Tibble with Admin names, year, age columns, datasource.
}
\description{
Load Somalia district population patch
}
