% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_dist_pop_data.R
\name{load_world_pop_patch}
\alias{load_world_pop_patch}
\title{Load 2015 World Pop data}
\usage{
load_world_pop_patch(
  world_pop_file_path = "GID/PEB/SIR/Data/pop/pop raw/csv files/adm2_2015_pop.csv",
  edav = TRUE
)
}
\arguments{
\item{world_pop_file_path}{\code{str} Path to the world pop file path.}

\item{edav}{\code{logical} Whether to use EDAV or not to load the file.}
}
\value{
\code{tibble} 2015 World Pop data.
}
\description{
World pop data obtained for 2015, to fill in missing data before 2016.
}
