% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_pop_outlier.R
\name{map_pop_outlier}
\alias{map_pop_outlier}
\title{Create a map to check for population outliers}
\usage{
map_pop_outlier(
  pop_w_outlier_cat,
  sf_long,
  pop_cat = "u15",
  ctry_name = NULL,
  who_region = NULL,
  year_end = lubridate::year(Sys.Date()),
  year_start = year_end - 3
)
}
\arguments{
\item{pop_w_outlier_cat}{\code{tibble} Output of \code{\link[=detect_processed_pop_outliers]{detect_processed_pop_outliers()}}.}

\item{sf_long}{\code{sf} Shapefile in long format.
Any of the sirfunctions::load_clean_*_sp(type = "long"), as long as the administrative level
is the same as the population file.}

\item{pop_cat}{\code{str} Population category to map. Either u15, u5, or totpop.}

\item{ctry_name}{\code{str} Optional country name.}

\item{who_region}{\code{str} Optional region name.}

\item{year_end}{\code{int} Year end. Defaults to current year.}

\item{year_start}{\code{int} Year start. Defaults to four years ago from the end year.}
}
\value{
\code{ggplot2} A map showing population count stability over the year
}
\description{
Maps the percentage of change in population over years
to check for outliers in the population data. This is a
diagnostic tool.
}
