% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pop_data_quality.R
\name{missing_required_agegroups}
\alias{missing_required_agegroups}
\title{Check if all required age groups are present}
\usage{
missing_required_agegroups(present_agegroups)
}
\arguments{
\item{present_agegroups}{\code{str} Character vector of age group codes.}
}
\value{
\code{str} \code{NA} if all present, otherwise a string of missing age groups.
}
\description{
Check if all required age groups are present
}
\keyword{internal}
