% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pop_data_quality.R
\name{outliers_for_guid_age}
\alias{outliers_for_guid_age}
\title{Orchestrates per guid + age group}
\usage{
outliers_for_guid_age(pop_data, guid_col, guid, age_group)
}
\arguments{
\item{pop_data}{Tibble containing population data.}

\item{guid_col}{String name of the GUID column to validate.}

\item{guid}{\code{str} GUID to search for.}

\item{age_group}{\code{str} Age group to search for.}
}
\value{
\code{tibble} Summary containing outliers per age group.
}
\description{
Filters and compute outliers for population values in the pop data
}
\keyword{internal}
