% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_dist_pop_data.R
\name{patch_polis_with_non_polis_pop}
\alias{patch_polis_with_non_polis_pop}
\title{Patch missing data in POLIS pop with non-polis data}
\usage{
patch_polis_with_non_polis_pop(polis_pop, non_polis_pop, patch_file)
}
\arguments{
\item{polis_pop}{\code{tibble} Population data from POLIS API.}

\item{non_polis_pop}{\code{tibble} Population data fron non-POLIS API sources.}

\item{patch_file}{\code{str} Name of the patch file datasource. Valid values include:
"JAMAL POP", "PATCH_PAKISTAN", "KENYA 2018 PATCH", "PATCH_SOMALIA".}
}
\value{
\code{tibble} Population data patched.
}
\description{
Patch missing data in POLIS pop with non-polis data
}
\keyword{internal}
