% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_ctry_pop_data.R
\name{process_ctry_pop_data}
\alias{process_ctry_pop_data}
\title{Build country population (Admin0) in wide format}
\usage{
process_ctry_pop_data(
  pop_data,
  pop_dir = "GID/PEB/SIR/Data/pop",
  ctry_file_path = "GID/PEB/SIR/Data/spatial/global.ctry.rds",
  growth_rate_file_path = file.path(pop_dir,
    "pop raw/WPP2024_GEN_F01_DEMOGRAPHIC_INDICATORS_COMPACT.xlsx"),
  output_dir = file.path(pop_dir, "processed_pop_file"),
  output_type = "parquet",
  edav = TRUE
)
}
\arguments{
\item{pop_data}{\code{tibble} Country population dataset pulled from the POLIS API.}

\item{pop_dir}{\code{str} Default directory to the population folder.}

\item{growth_rate_file_path}{\code{str} File path to the growth rate Excel file.}

\item{output_dir}{\code{str} Where to output the cleaned district population dataset.}

\item{output_type}{\code{str} Output types. Valid values are 'rds', 'csv', and 'parquet'. Default
is \code{rds}.}

\item{edav}{\code{logical} Whether to use EDAV or not to load the file.}
}
\value{
\code{tibble} Cleaned country population.
}
\description{
Combines POLIS + patches + Jamal; joins to country-year shapes; deduplicates;
fills datasource across gaps within GUID; fills missing values using growth rates.
Aggregates remaining country-level gaps by summing all province totals per country.
}
