% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_pop_data.R
\name{remove_forward_fill_non_polis}
\alias{remove_forward_fill_non_polis}
\title{Remove forward-filled values (non-POLIS only)}
\usage{
remove_forward_fill_non_polis(non_polis_pop)
}
\arguments{
\item{non_polis_pop}{Tibble with ADM2_GUID, year, age columns, datasource.}
}
\value{
Tibble with forward-filled values set to NA.
}
\description{
Remove forward-filled values (non-POLIS only)
}
