% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_forward_fill_polis_pop.R
\name{remove_forward_fill_polis_pop}
\alias{remove_forward_fill_polis_pop}
\title{Remove Forward-Filled POLIS Population Values}
\usage{
remove_forward_fill_polis_pop(pop_data)
}
\arguments{
\item{pop_data}{\code{tibble} Population data.}
}
\value{
\code{tibble} Population data with forward-filled values set to NA.
}
\description{
Identifies forward-filled population records and sets them to NA.
}
\examples{
\dontrun{
remove_forward_fill_polis_pop(pop_data)
}

}
