% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survpopdata-package.R
\docType{package}
\name{survpopdata-package}
\alias{survpopdata}
\alias{survpopdata-package}
\title{survpopdata: Create Cleaned Population Datasets Used In GPEI Analyses}
\description{
The package aims to clean the population datasets used in multiple GPEI analyses, from the ADM0 through the ADM2 levels.
}
\author{
\strong{Maintainer}: Mervin Keith Cuadera \email{xrg9@cdc.gov} (\href{https://orcid.org/0000-0003-4898-2659}{ORCID})

Authors:
\itemize{
  \item Philip Bammeke \email{philip.bammeke@gmail.com}
}

}
\keyword{internal}
