% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pop_data_quality.R
\name{create_parent_predicate}
\alias{create_parent_predicate}
\title{Create unique parent predicate per admin GUID for assertr validation}
\usage{
create_parent_predicate(pop_data, guid_col)
}
\arguments{
\item{pop_data}{Tibble containing population data.}

\item{guid_col}{String name of the GUID column to validate.}
}
\value{
A predicate function for use with assertr.
}
\description{
Predicate that checks whether each admin GUID maps to a single parent admin.
}
\details{
For "Admin2GUID", checks uniqueness against Admin1GUID and Admin0GUID.
For "Admin1GUID", checks uniqueness against Admin0GUID.
}
\keyword{internal}
