% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_dist_pop_data.R
\name{load_jamal_pop}
\alias{load_jamal_pop}
\title{Load Jamal district under-15 population}
\usage{
load_jamal_pop(
  jamal_pop_file_path = "GID/PEB/SIR/Data/pop/pop raw/csv files/POPU15.csv",
  edav = TRUE
)
}
\arguments{
\item{jamal_pop_file_path}{Path to Jamal file.}

\item{edav}{\code{logical} Whether to use EDAV or not to load the file.}
}
\value{
\code{tibble} Population data with Admin names, year, age columns, datasource.
}
\description{
Load Jamal district under-15 population
}
