% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_pop_data.R
\name{load_somalia_patch}
\alias{load_somalia_patch}
\title{Load Somalia district population patch}
\usage{
load_somalia_patch(
  somalia_2022_file_path = "GID/PEB/SIR/Data/pop/pop raw/csv files/AFPPOP_22.csv",
  somalia_2023_file_path = "GID/PEB/SIR/Data/pop/pop raw/csv files/AFPPOP_23.csv",
  somalia_2024_file_path = "GID/PEB/SIR/Data/pop/pop raw/csv files/AFPPOP_24.csv",
  edav = TRUE
)
}
\arguments{
\item{somalia_2022_file_path}{Path to Somalia 2022 file.}

\item{somalia_2023_file_path}{Path to Somalia 2023 file.}

\item{somalia_2024_file_path}{Path to Somalia 2024 file.}
}
\value{
Tibble with Admin names, year, age columns, datasource.
}
\description{
Load Somalia district population patch
}
