% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_dist_pop_data.R
\name{process_dist_pop_data}
\alias{process_dist_pop_data}
\title{Build district population}
\usage{
process_dist_pop_data(
  pop_data,
  pop_dir = "GID/PEB/SIR/Data/pop",
  dist_file_path = "GID/PEB/SIR/Data/spatial/global.dist.rds",
  growth_rate_file_path = file.path(pop_dir,
    "pop raw/WPP2024_GEN_F01_DEMOGRAPHIC_INDICATORS_COMPACT.xlsx"),
  pakistan_file_path = file.path(pop_dir,
    "pop raw/csv files/2022_2023 Population Pakistan.csv"),
  somalia_2022_file_path = file.path(pop_dir, "pop raw/csv files/AFPPOP_22.csv"),
  somalia_2023_file_path = file.path(pop_dir, "pop raw/csv files/AFPPOP_23.csv"),
  somalia_2024_file_path = file.path(pop_dir, "pop raw/csv files/AFPPOP_24.csv"),
  kenya_file_path = file.path(pop_dir, "pop raw/csv files/Kenya_SubCounty_pop_2018.csv"),
  jamal_pop_file_path = file.path(pop_dir, "pop raw/csv files/POPU15.csv"),
  world_pop_file_path = file.path(pop_dir, "pop raw/csv files/adm2_2015_pop.csv"),
  output_dir = file.path(pop_dir, "processed_pop_file"),
  output_type = "parquet",
  edav = TRUE
)
}
\arguments{
\item{pop_data}{\code{tibble} District population dataset pulled from the POLIS API.}

\item{pop_dir}{\code{str} Default directory to the population folder.}

\item{dist_file_path}{\code{str} File path to the global district shapefile.}

\item{growth_rate_file_path}{\code{str} File path to the growth rate Excel file.}

\item{pakistan_file_path}{Path to Pakistan patch file.}

\item{somalia_2022_file_path}{Path to Somalia 2022 file.}

\item{somalia_2023_file_path}{Path to Somalia 2023 file.}

\item{somalia_2024_file_path}{Path to Somalia 2024 file.}

\item{kenya_file_path}{Path to Kenya 2018 file.}

\item{jamal_pop_file_path}{Path to Jamal file.}

\item{world_pop_file_path}{\code{str} Path to the world pop file path.}

\item{output_dir}{\code{str} Where to output the cleaned district population dataset.}

\item{output_type}{\code{str} Output types. Valid values are 'rds', 'csv', and 'parquet'. Default
is \code{rds}.}

\item{edav}{\code{logical} Whether to use EDAV or not to load the file.}
}
\value{
\code{tibble} Cleaned district population dataset.
}
\description{
Cleans the district level population data from the POLIS API and fill in
gaps in population counts using several patch files and application of growth rates.
}
\examples{
\dontrun{
pop_data <- load_polis_pop("dist")
dist_pop <- process_dist_pop_data(pop_data)
}
}
