% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_prov_pop_data.R
\name{process_prov_pop_data}
\alias{process_prov_pop_data}
\title{Build province population (Admin1) in wide format}
\usage{
process_prov_pop_data(
  pop_data,
  pop_dir = "GID/PEB/SIR/Data/pop",
  dist_pop_file_path = file.path(pop_dir, "processed_pop_file", "dist.pop.long.parquet"),
  prov_file_path = "GID/PEB/SIR/Data/spatial/global.prov.rds",
  growth_rate_file_path = file.path(pop_dir,
    "pop raw/WPP2024_GEN_F01_DEMOGRAPHIC_INDICATORS_COMPACT.xlsx"),
  output_dir = file.path(pop_dir, "processed_pop_file"),
  output_type = "parquet",
  edav = TRUE
)
}
\arguments{
\item{pop_data}{\code{tibble} Province population dataset pulled from the POLIS API.}

\item{pop_dir}{\code{str} Default directory to the population folder.}

\item{dist_pop_file_path}{\code{str} File path to the cleaned district pop file.}

\item{prov_file_path}{\code{str} File path to the global province shapefile.}

\item{growth_rate_file_path}{\code{str} File path to the growth rate Excel file.}

\item{output_dir}{\code{str} File path to the output directory.}

\item{output_type}{\code{str} How the population file should be outputted.}

\item{edav}{\code{logical} Whether files are on EDAV. Defaults to \code{TRUE}.}
}
\value{
\code{tibble} Tibble with GUID + province-year rows and final output naming.
}
\description{
Combines POLIS + patches + Jamal; joins to province-year shapes; deduplicates;
fills datasource across gaps within GUID; fills missing values using growth rates.
Aggregates remaining province-level gaps by summing all district totals per province.
}
\examples{
\dontrun{
prov_pop_data(pop_data, output_file = "Data/pop/prov_pop_admin1.rds")
}
}
